Use some ES6 features to make code more structured.
No functional changes, except a few API changes.

Removed the dependency on jQuery in library code. This breaks the API in a
couple places, such as `initscr()`.

Switched from Grunt to Webpack for building, and added Babel to the
toolchain. Also switched linters.

Here are the main ES6-related changes:
* `import` and `export` directives
* `const` or `let` to replace `var` in some places (needs more work)
* change `getch(callback)` to `getch()` that returns a `Promise`
* use the `class` keyword to declare types